refactor(auth): unify reset password form into reusable component
- Extract ResetPasswordForm from user and admin apps into libs/ui library
- Simplify reset password pages to import and use ResetPasswordForm component with props
- Improve type safety for forgot password mutation and error handling
- Enhance email service with validation for SMTP port and URL construction using URL API
- Update password reset response message for privacy
- Use Suspense fallback in reset password pages for loading state
- Preserve UI and toast notifications behavior in reusable ResetPasswordForm component
