+++ libs/ui/src/index.ts
@@ -11,3 +11,4 @@ export { AdminSidebar } from './components/AdminSidebar';
 export { DashboardChart } from './components/DashboardChart';
 export { ReactQueryProvider } from './components/react-query-provider';
 export { Pagination } from './lib/Pagination';
+export { ResetPasswordForm } from './components/auth/ResetPasswordForm';
